[Feature] #234 player modularity : converting mp3 to wav using param
Done : 
-> Travis back using python requirements instead of setup.py
-> convert_to_wav param
