+++ .travis.yml
@@ -21,8 +21,7 @@ before_install:
 - sudo python get-pip.py
 
 install:
-  - sudo python setup.py install
-  # - pip install -r install/files/python_requirements.txt
+  - pip install -r install/files/python_requirements.txt
   - pip install python2-pythondialog || pip install pythondialog
   - pip install coveralls
