+++ kalliope/settings.yml
@@ -37,15 +37,15 @@ speech_to_text:
   - google:
       language: "fr-FR"
   - wit:
-      key: "B5JI3YUSLYOYWNIDBINBVM34XUODME2K"
+      key: "fakekey" # example : "B5JI3YUSLYOYWNIDBINBVM34XUODME2K"
   - bing: # API not working : credential fails ..
-      key: "9e48ddaf75904838bedc11aea6b36fb0"
+      key: "fakekey" # example : "9e48ddaf75904838bedc11aea6b36fb0"
   - apiai:
-      key: "e0cbff154af44944a6b9f82c0668b527"
+      key: "fakekey" # example : "e0cbff154af44944a6b9f82c0668b527"
       language: "fr"
   - houndify:
-      key: "7zj90T7qAV74OYXk4X4vI2Xhk7wPsJu4aEZ0G5Ll-BMmV1JGtFpCxtSH9SmTY4G3bpEJ7a5y_GTQid-CAKI6vw=="
-      client_id: "lN4JXeaSticbSo9-llczbA=="
+      key: "fakekey" # example : "7zj90T7qAV74OYXk4X4vI2Xhk7wPsJu4aEZ0G5Ll-BMmV1JGtFpCxtSH9SmTY4G3bpEJ7a5y_GTQid-CAKI6vw=="
+      client_id: "fakeclientid" # example : "lN4JXeaSticbSo9-llczbA=="
   #- cmusphinx
 
 
@@ -60,10 +60,9 @@ cache_path: "/tmp/kalliope_tts_cache"
 # Text to Speech engines configuration
 # Available engine are:
 # - pico2wave
-# - acapela
-# - pico2wave
-# - googletts
-# - voicerss
+# - acapela     # MP3
+# - googletts   # MP3
+# - voicerss    # MP3
 text_to_speech:
   - pico2wave:
       language: "fr-FR"
@@ -88,15 +87,18 @@ default_player: "mplayer"
 # players configuration
 # Available engine are:
 # - mplayer
-# - pyalsaaudio
-# - pyaudioplayer
-# - sounddeviceplayer
+# - pyalsaaudio       # no mp3
+# - pyaudioplayer     # no mp3
+# - sounddeviceplayer # no mp3
 players:
   - mplayer: {}
   - pyalsaaudio:
      device: "default"
-  - pyaudioplayer: {}
-  - sounddeviceplayer: {}
+     convert_to_wav: True
+  - pyaudioplayer:
+     convert_to_wav: True
+  - sounddeviceplayer:
+     convert_to_wav: True
 
 # ---------------------------
 # Wake up answers
